Issue 5636/4: int->long for result of scm_ilength ()
to fix type-conversion warnings

+++ lily/dispatcher.cc
@@ -86,7 +86,7 @@ Dispatcher::dispatch (SCM sev)
 #if 0
   bool sent = false;
 #endif
-  int num_classes = scm_ilength (class_list);
+  long num_classes = scm_ilength (class_list);
 
   /*
     For each event class there is a list of listeners, which is
